Route nested and forms modified

+++ app/Http/Requests/UserGraduationRequest.php
@@ -25,9 +25,9 @@ class UserGraduationRequest extends FormRequest
     {
         return [
             'is_graduated'                  => ['required', 'boolean'],
-            'year'                          => ['required', 'numeric', 'min:0', 'max:2100'],
-            'educational_institution_id'    => ['required', 'numeric', 'exists:educational_institutions,id'],
-            'academic_program_id'           => ['required', 'numeric', 'exists:academic_programs,id']
+            'year'                          => ['required', 'integer', 'min:0', 'max:2100'],
+            'educational_institution_id'    => ['required', 'integer', 'exists:educational_institutions,id'],
+            'academic_program_id'           => ['required', 'integer', 'exists:academic_programs,id']
         ];
     }
 }
